requirements.txt
streamlit>=1.32.0
python-docx>=1.1.0
ollama>=0.1.6
numpy>=1.26.4
scikit-learn>=1.4.1
tqdm>=4.66.0